chore: Refatoring project
